Complete pipeline for building dev releases
This pipeline represents the build process for a single Bosh release
using the pullrequest-resource[1]. During development variables were
passed in using an external file and the `--load-vars-from` option.

The Bosh releases used were forks of the paas-example-boshrelease[2]
and were held in private repositories which we generated tokens for.
A real implementation would likely use a bot account with its own keys.

It detects pull requests raised against the Bosh release repository
and triggers a job which creates a dev Bosh release and uploads it
to an S3 bucket (created manually). The task then outputs the necessary
YAML to add to the Bosh manifest for Bosh to be able to use the release.

[1] https://github.com/jtarchie/pullrequest-resource
[2] https://github.com/alphagov/paas-example-boshrelease