Initial version of pipeline template
Initially we had not set the `REGION` parameter and the job did not
fail, despite having `set -u`. We therefore chose to use a bash shell
so that errors like this will cause the build to error in future
(the busybox sh implementation doesn't exit non-zero for unbound
variables in a HEREDOC, even with the `-u` parameter).