Use version:every for the pullrequest resource.
This uses the new behaviour in concourse 1.2+ that causes the get to
iterate through every version returned instead of picking the latest by
default.

This is desireable with the PR resource because the list of
pull-requests don't map well to versions, and the PR resource has had to
implement logic to only return the latest unseen version in the check.
It does this by looking at the github status events set on the PR
commits. This has the problen that if multiple concourse deployments are
pointing at the same repo, they condlict with each other, ane only one
of them will see a given version. (see [1] for more details). This
behafiour is turned off by setting `every: true` in the resource config.

[1]https://github.com/jtarchie/pullrequest-resource#behavior