Add ability to upload CF CLI secrets
The release CI requires credentials to deploy to the target environment.
A user was created manually in production with enough permissions to
deploy, the credentials were stored in paas-pass.
This make target allows to upload manually the credentials to S3 so they
can be retrieved later.
For dev environments, the user and password can be passed by environment
variables.