Generic application deployment pipeline
Similarly to bosh releases, the setup pipeline creates pipelines for
each application we wish to deploy.
The parameters for each application deployment are stored in
app-deploy.d and the script build-app-deployment-pipelines.sh reads them
to deploy each app.
The pipeline is composed of 2 jobs and 3 tasks:

* test: calls the release/test script in the app if present
* build: calls the release/build script in the app if present and stores
the files for push
* push: calls the release/push script in the app if present. If not, it
deploys the application using zero downtime.