Fix when CF_PASWORD is empty
In Dev we may want to test bosh releases and not care about using the
cf api, so the CF_USER and CF_PASSWORD may not be set.
This avoids breaking in Dev by setting an empty string.

Also, CF_USER and CF_PASSWORD are set in environment.sh so we don't need
to retrieve them in the setup script.