Fix path to CDN broker Bosh release repository
This has changed now that we have a separate repository for the Bosh
release, which submodules in the CDN broker.

This was done because it allows us to contribute changes upstream and
it stops us breaking Go import paths. We still need to fix the three-
pull-request dance we do when developing Bosh releases, but that can
be addressed as a different piece of work (preferably on the RDS broker,
since it is more stable and is not a fork).