Change pipeline pausing in dev
It has long confused me that you need to manually run the `setup` pipeline
when you create a Build Concourse in a dev environment, but then *all* of
the other pipelines run and start submitting build results back to GitHub.

I think it makes more sense to automatically trigger the `setup` pipeline
but pause all of the release pipelines. Then you can unpause the release
that you intend to work on.

I've set the `setup` and `destroy` pipelines to explicitly not be paused.