Rename release pipelines to have -release suffix
We want to introduce new pipelines to test pull requests on the code repos,
e.g. paas-rds-broker rather than paas-rds-broker-boshrelease. The existing
naming makes this confusing.

We need to use a command with a FIXME comment to rename the pipelines
in-place so that the build history is kept. This should be a noop on the 2nd
run. Then the `rename-pipeline` command can be removed.