Handle concourse password changes in self-update-pipelines
The self-update-pipelines currently will error if the concourse password
is changed. This is because the password is passed around in a loop, and
therefore the new one isn't pulled from the S3 bucket. Breaking this
loop means that the password will be pulled from S3 each time, and
therefore will be updated.