Deploys apps in serial
Because I don't know what Cloud Foundry will do if two commits land on the
master branch in quick succession and result in two parallel deployments
happening at the same time.

I noticed this because app deployments are currently broken by credential
rotation in Pivotal story #146152015 and this has caused several builds to
get stuck in the pending state.

It is still safe to run the test task in parallel.