Fix quoting of passwords in YAML and scripts.
We discovered that a password starting with '!!' caused a number of
problems due to insufficient quoting:

Firstly, ! has special meaning in YAML relating to field tags. Values
starting with this therefore need to be quoted.

Secondly, !! has special meaning in some shells. The output from the
environment.sh script is eval'd, which was causing breakage. The values
have to be single quoted to avoid this.