Make pipecleaner handle external task files better
This is a partial revert of e47d67b with a different implementation. It
checks whether the file exists with the resource name/path stripped of the
beginning. We work around this in alphagov/paas-cf by executing `cd ..`
before `pipecleaner`, but this is nicer.

Ignore the unused resources check when external task files (defined in
another repo) have been used because we can't see when `inputs` they
have defined.

This was a problem for alphagov/paas-compose-broker integration tests
which use a `secrets` resource that, unlike the `pr` resource, isn't use in
`on_success` or `on_failure`.

It's possible to fetch a resource using a different name, e.g.:

    - get: pr
      resource: release-repository

Without this pipecleaner complained that the resource was unused.