Add paas-submit app
This app contains forms for the product page and a way of submitting
their contents to the DeskPro API to raise tickets.

We use the Docker image of the paas-product-page because the `release/test`
and `release/build` steps are the only tasks that use the image, and they
are both no-ops for this app (the tests will run in Travis and the Ruby
buildpack will handle all of the building).