Update README
The setup instructions were missing details of the latest secrets that
need to be uploaded to the state bucket (Compose API key and now the
DeskPro API key).

I have merged the requirements and setup sections because I believe it makes
it easier to read and follow sequentially.

Also, security groups prevent access to dev Cloud Foundry deployments
from a build CI environment, so we needed to document the workaround.