Add route-registrar pipeline
As part of #151322324, we've decided that we would like to fork the
upstream route-registrar repo in order to extend its functionality to
support the TLS port. This work has already been done but not merged in.
See: cloudfoundry/route-registrar#18.

We intend to go back into using the upstream release as soon as there is
a similar fix merged in, making this commit obsolete.