Pull in pipecleaner changes from paas-cf
This pulls in the latest changes to pipecleaner from paas-cf -
specifically, the dummy variables that it inserts before shellchecking
have had a date interpolation added so that shellcheck doesn't think
they're safe strings.

This retains the changes that have been added in da7900e to improve the
handling of external tasks.