Point to a secret file
As it turns out, the integration tests require some secrets in order to
connect to the compose API and run checks before creating a tag release.

This parameter should allow the script to obtain the path to it and fill
in the gaps.

Some of the parameters that have been referenced thus far, are not
actually being used. Therefore we took the chance to remove them.