Flag to specify docker image tag for app deploys
Until now, the app deploy config files in `app-deploy.d` have not been
able to specify a tag for the docker image they are using for their
`test` and `build` steps. This is inconvenient: the only way to
specify a tag has been to manually edit an app's deployment pipeline.

This commit adds a `APP_DOCKER_IMAGE_TAG` flag that can be specified
in `app-deploy.d` config files. This flag default to `latest`, which
was the existing, implicit default.