Explicitly set release name when creating release (and 2)
Note: Same than fa60d8512699a8b7cfce21f9912dbf353c4fadbb but for the
master branch

After upgrading the bosh cli to v2 as part of the PR
alphagov/paas-release-ci#51 we dropped the option `--name`,
as it was not docummented online [1].

But after doing that, releases like datadog-for-cloudfoundry would
get a different name than before (e.g. paas-datadog-for-cloudfoundry
instead of datadog-for-cloudfoundry.

We add again the explicit option `--name` to fix this.

[1] https://bosh.io/docs/cli-v2.html