Move paas-product to a plain-pipeline
The current app-deploy.d feels complicated, as involves multiple
files across different projects in different languages, parsing
and passing variables explicitly across.

It couples both paas-release-ci with the other application projects,
as one must set the right variables consumed by the later. Any change
must be done in sync across all the projects.

There is logic of parsing the secrets and rendering the manifest
from files in the paas-product-page, which makes it difficult to
change later.

Specifically, in our case, to add the ZenDesk secrets would require
several PRs and changes.

Instead, we propose a simpler approach of keeping a plain pipeline
in this project with all the logic required. We use spruce[1]
to render a manifest and inject any secret

[1] https://github.com/alphagov/paas-docker-cloudfoundry-tools/pull/124