Publish paas-team-manual to Github Pages
Add a pipeline that will build the paas-team-manual and push it
as a github page.

Target URL: https://alphagov.github.io/paas-team-manual/

For simplicity we use the `rake publish` task directly,
which requires run  git to create a branch for gh-pages and push it.

The job must:
 - Install the dependencies for ruby, git and ssh
 - Configure git+ssh: set user, ssh key, rewrite origin url...

We use ruby based on debian as the task image because it is easier to
get the ruby dependencies working fine. Specifically, it is easier to
install the libv8 library for the js engine.

We use the ci_build_tag_key configured in setup.sh

[1] https://pages.github.com/