Create individual pipelines for tech-docs and rubbernecker
We tried to handle the test/build/deploy process for all apps in a common
pipeline but it resulted in multiple ifs and smaller hacks to handle all use
cases. We decided that having individual pipelines for the applications will
make modifications easier and result in simpler pipelines.

See [1] for more details.

[1] https://github.com/alphagov/paas-release-ci/pull/59/commits/8e991992da6506b8e3a45fda04b4d6067145cf1a