Build our own capi-release
We are having trouble running the capi-release[1]. One solution seems
to be rebuild our own release tarball, without changes.

Other option is to use precompile releases, but as commented
by @jpluscplusm

"our IA "2 eyes" continuous deployment "agreement" is based on the
potential for us to examine the code we're running in prod, and how that
interacts with upstream's CI process and security guarantees. In the
case of An Issue, could we 100% point at the code we're running?"

Because that, we decided to use our build release from code.

[1] https://cloudfoundry.slack.com/archives/C07C04W4Q/p1541779348165300
[2] https://github.com/cloudfoundry/cf-deployment/issues/663