Reduce the check interval for all resources
We are having load issues on our build CI, as all the resources
are checking for new versions across all the pipelines.

But we only want to check for those resources that trigger pipelines,
as the rest of the resources are checked as soon as each job is
triggered.

We will increase `check_every` to a big period for all non triggering
resources. We pick `24h`, to effectivelly disable the checks
on these resources.

For the resources that actually trigger the pipeline, we will keep
the default of 1m, but set it explicitelly to 1m.

More info:

[1] https://github.com/concourse/concourse/issues/1842
[2] https://github.com/concourse/concourse/issues/1445