Add Slack notifications when deploys from CI fail
Our "release CI" automation [1] builds things (e.g., BOSH Releases.)
It also directly deploys some software:

  - paas-tech-docs
  - paas-product-page
  - paas-team-manual
  - rubbernecker
  - paas-hackmd

People rarely look at this Concourse and so failures aren't noticed.
The deployment pipeline for paas-tech-docs has been failing a lot
recently [2]. This causes changes in a merged Pull Request to not be
deployed, and this in turn may not be noticed.

This PR adds a Slack notification when any of the directly-deployed
things fail. This will make failures more visible, which in turn may
lead to better solutions.

[1] https://concourse.build.ci.cloudpipeline.digital
[2] Runs 186, 188, 189, 190, 191, 194, 196, 197, 198, 199, 201, 204 and 205 of https://concourse.build.ci.cloudpipeline.digital/teams/main/pipelines/paas-tech-docs/jobs/test/builds/205 failed, delaying some releases by as much as two weeks!