Build Docker Containers In Concourse
This adds two pipelines that build docker containers to our build concourse.
This is so we can migrate from Docker Hub's Automated Builds.

`build-docker-containers.yml` is triggered when a branch is merged to
master. `build-docker-pull-requests.yml` is triggered when a pull request is
raised and tags the image with pr-build, pr-branch-name and hash. `pr-build` is
specified because the registry-image-resource defaults to latest.

In order to push the images to Docker Hub, this adds the username/password from
the password store.