Pipeline must install paas-tech-docs dependencies
Since merging a gem upgrade [1] the `paas-tech-docs` pipelines on the
build concourse have been failing with errors like the following:

  bundle exec middleman build
  Could not find rack-2.1.1 in any of the sources
  Run `bundle install` to install missing gems.

It turns out that `paas-tech-docs` has a `dependencies` Make target
that the pipeline wasn't invoking. I checked locally and there are no
`Gemfile.lock` changes made when invoking it, so I don't think that
it's a problem with `paas-tech-docs`. I'm *guessing* that this worked
until now because the Gems being used happened to be in the Docker
container.

[1] https://github.com/alphagov/paas-tech-docs/pull/288