Add credhub-shell script
In order to interact with Credhub, we need to open an SSH tunnel to the
environment's Bosh director. We have a good script for this in `paas-cf`, so
I've copied it over to this repository and made some small tweaks to the
output.

It can now be used via Make: `make {env} credhub`