Setup a Session Secret
We're going to be generating a new secret everytime we run
`make upload-hackmd-secrets`, meaning we have an easy way to rotate the
secret, but also ensure that the secret persist in the bucket for long
enough, to be pulled out and used by the application itself, to use for
session encryption.

This should ensure, that restarting or re-scheduling the application
will not log everyone out at all times. It also means, we can have more
than a single instance of CodiMD running at once, increasing it's
reliability.