prs: update pending step in separate context
from the docs:

---
When specifying skip_download the pull request volume mounted to
subsequent tasks will be empty, which is a problem when you set e.g. the
pending status before running the actual tests. The workaround for this
is to use an alias for the put (see
https://github.com/telia-oss/github-pr-resource/issues/32 for more
details). Example here:

put: update-status <-- Use an alias for the pull-request resource
resource: pull-request
params:
    path: pull-request
    status: pending
get_params: {skip_download: true}
---

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>