Build docker images that mirror DockerHub
Until now we have relied on third-party DockerHub images, such as
`ruby:2.7-slim`. But DockerHub's rate limiting means that image is no
longer so reliable. We want to switch to use Github Container
Registry (GHCR) in production.

There isn't an equivalent image on GHCR, so we're going to make one.
This commit adds a pipeline job that will build us a Ruby container
and upload it to:

  - `governmentpaas/ruby` on DockerHub
  - `ghcr.io/alphagov/paas/ruby`, for production use

We also do the same for `concourse-pool-resource` and `alpine`. See
`paas-docker-cloudfoundry-tools` for the `Dockerfile`s used.