Update tech-docs deploy manifest
Remove `path` and `buildpack` attributes

In https://github.com/alphagov/paas-tech-docs/pull/383 we're adding a `Staticfile` to define error page templates.

There appears to be a bug where if the manifest file contains the path and buildpack attributes and the repo contains a Staticfile,
any setting in the Staticfile doesn't get applied.